Start integrating with third party apps
Create an interface to contain the methods the classes should implement.
Create two classes to implement that interface, one for Alphavantage and another for Marketstack.
Create route to get the latest data available from marketstack for a given company.
Persist this data in the database.

TODO: Instead of doing this route sync, use a message queue to process it, or use the scheduler to run daily.